fix: move alphafold loading step into importing and some minor fixes
