+++ backend/protzilla/all_steps.py
@@ -14,6 +14,7 @@ _forward_mapping = [
     importing.EvidenceImport,
     importing.ExampleDatasetImport,
     importing.FastaImport,
+    importing.AlphaFoldPredictionLoad,
     data_preprocessing.FilterProteinsBySamplesMissing,
     data_preprocessing.FilterProteinsBySilacRatios,
     data_preprocessing.FilterByProteinsCount,
@@ -65,7 +66,6 @@ _forward_mapping = [
     data_analysis.PTMOverviewVisualization,
     data_analysis.PTMBarVisualization,
     data_analysis.PTMDetailsVisualization,
-    data_analysis.AlphaFoldPredictionLoad,
     data_preprocessing.ImputationByMinPerSample,
     data_integration.EnrichmentAnalysisGOAnalysisWithString,
     data_integration.EnrichmentAnalysisGOAnalysisWithEnrichr,
